+++ index.html
@@ -233,7 +233,6 @@
         <h2 class="data-error__title">Не удалось загрузить данные</h2>
       </section>
     </template>
-    <script src="functions.js" type="module"></script>
-    <script src="js/main.ts" type="module"></script>
+    <script src="./js/main.ts"></script>
 </body>
 </html>
